Fix CI for ubuntu-24.04 runners
Github have just updated ubuntu-latest runners to use the new
ubuntu-24.04 images again, with their reduced set of software
pre-installed. The CI on this repo has started failing as a result,
because it relies on sbt being installed by default.

To fix this, this commit replaces the use of actions/setup-java with our
own action
[guardian/setup-scala](https://github.com/guardian/setup-scala), which
we’ve been using for this purpose. It installs both Java and sbt, and
configures the Java version from the .tool-versions file in the
repository.

This does mean that this change moves the CI from using Java 11 to Java
21: hopefully this is not a problem.